+++ R/lifecycle.R
@@ -1,18 +1,13 @@
 #' Use lifecycle badges
 #'
 #' @description
+#' This helper add lifecycle as a dependency, copies the lifecycle badges into
+#' the `man/figures` folder, and reminds you how to use the badge syntax.
 #'
-#' This helper copies the lifecycle badges in to the `man/figures`
-#' folder of your package. It also reminds you of the syntax to use
-#' them in the documentation of individual functions or arguments.
-#'
-#' See the [getting started
-#' vignette](https://lifecycle.r-lib.org/articles/lifecycle.html) of the
-#' lifecycle package.
+#' Learn more at <https://lifecycle.r-lib.org/articles/communicate.html>
 #'
 #' @seealso [use_lifecycle_badge()] to signal the [global lifecycle
 #'   stage](https://www.tidyverse.org/lifecycle/) of your package as a whole.
-#'
 #' @export
 use_lifecycle <- function() {
   check_is_package("use_lifecycle()")
@@ -21,6 +16,8 @@ use_lifecycle <- function() {
     ui_stop("
       Turn on roxygen2 markdown support {ui_code('use_roxygen_md()')}")
   }
+
+  use_package("lifecycle")
 
   dest_dir <- proj_path("man", "figures")
   create_directory(dest_dir)
@@ -33,15 +30,9 @@ use_lifecycle <- function() {
 
   ui_todo(c(
     "Add badges in documentation topics by inserting one of:",
-    "- `r lifecycle::badge('experimental')`",
-    "- `r lifecycle::badge('superseded')`",
-    "- `r lifecycle::badge('questioning')`",
-    "- `r lifecycle::badge('deprecated')`"
-  ))
-
-  ui_todo(c(
-    "If you want to use functions like `lifecycle::deprecate_soft()` in your package:",
-    "- `usethis::use_package(\"lifecycle\")`"
+    "#' `r lifecycle::badge('experimental')`",
+    "#' `r lifecycle::badge('superseded')`",
+    "#' `r lifecycle::badge('deprecated')`"
   ))
 
   invisible(TRUE)